LangGraphIntro/requirements.txt
langgraph>=0.0.20
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-core>=0.1.0
python-dotenv>=1.0.0
ipython>=8.0.0
openai>=1.0.0
